requirements.txt
pandas==2.0.3
numpy==1.24.3
streamlit==1.31.1
Pillow==10.0.1
catboost==1.2.3
scikit-learn==1.4.1.post1
pickle-mixin==1.0.2